% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cellrank_params.R
\name{check_cellrank_params}
\alias{check_cellrank_params}
\title{A check function for NetID}
\usage{
check_cellrank_params(params)
}
\arguments{
\item{mode}{the velocity mode, default: stochastic}

\item{weight_connectivities}{default: 0.2}

\item{ncores}{default: 6}

\item{plot}{default: TRUE}

\item{tolerance}{default: 1e-6}
}
\description{
a check function for FateDynamic function.
FateDynamic have provided cellrank_params, user could specific the parameters the want, the parameters are listed as follow
}
