% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_global_params.R
\name{check_global_params}
\alias{check_global_params}
\title{A Function to Validate NetID Parameters}
\usage{
check_global_params(params)
}
\arguments{
\item{min_counts}{Default: 20.}

\item{nhvgs}{Default: 2000.}

\item{npcs}{Default: 30.}

\item{knn}{Default: 30.}

\item{velo}{If the velocity model (cellrank) is run, default: TRUE.}

\item{fate_predict}{Default: TRUE.}

\item{cluster_label}{Specify the attribute in the SingleCellExperiment object that contains cluster label information.}
}
\description{
This function validates parameters for the FateDynamic function.
FateDynamic provides global_params; users can specify the desired parameters, which are listed as follows:
}
