% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_netID_params.R
\name{check_netID_params}
\alias{check_netID_params}
\title{A check function for NetID}
\usage{
check_netID_params(params)
}
\arguments{
\item{var}{if using variable gene to calculate principal components, used by geosketch method, default: FALSE}

\item{sampled_cells}{the barcode or ID of sampled cells}

\item{sketch.method}{perform sketching sampling on single cell datasets, "geosketch" or "SeuratSketching"}

\item{ndim}{dimensions of PCs, used by geosketch method, Default: 30}

\item{n_cell}{the number of sampled cells, default: 500}

\item{Threshold_Num}{the minimum nn of each seed cells after assignments, default: 2}

\item{normalize}{if perform normalization to the count matrix, default: FALSE}

\item{prior_net}{if a binary matrix indicate the prior knowledge of gene regulation, row is the regulator, column is the target.}
}
\description{
a check function for RunNetID function.
RunNetID have provided netID_params, user could specific the parameters the want, the parameters are listed as follow
}
