% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_palantir_params.R
\name{check_palantir_params}
\alias{check_palantir_params}
\title{A Function to Validate NetID Parameters}
\usage{
check_palantir_params(params)
}
\arguments{
\item{ndcs}{The number of diffusion components. Default: 10.}

\item{start_cell}{The barcode or ID of the root cell. Default: NULL.}

\item{terminal_state}{The terminal cell type labels.}

\item{nwps}{The number for waypoint sampling. Default: 500.}

\item{method}{Specify a method to find the root cell. Three options include: "scent", "cytotrace", and "markergene".}

\item{root_gene}{If choosing "markergene", please specify the ID of the gene.}

\item{plot}{If generating the plots in the working directory.}
}
\description{
This function validates parameters for the FateDynamic function.
FateDynamic provides palantir_params; users can specify the desired parameters, which are listed as follows:
}
