% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FateCausal.R
\name{FateCausal}
\alias{FateCausal}
\title{Run NetID model}
\usage{
FateCausal(
  dyn.out,
  velo_infor = FALSE,
  L = 30,
  alpha = 0,
  lambda = 100,
  cutoff = 0.25,
  weight = 0.2,
  redirected = FALSE,
  fate_method = "palantir",
  aggre_method = "RobustRankAggreg",
  restart = FALSE,
  n.cores = 8,
  work_dir = NULL
)
}
\arguments{
\item{dyn.out}{The output of NetID.}

\item{velo_infor}{if use velocity information to infer gene regulatory network
Default: FALSE}

\item{L}{The number of lags to utilize
Default: 30}

\item{alpha}{use ridger (0) or lasso regression (1)
Default: 0}

\item{lambda}{regularization parameters
Default: 100}

\item{cutoff}{when use velocity aid GRN construction, FateCausal would convert the scaled velocity value > |cutoff| or < -|cutoff| into 1 and 0 to denote induction or repression. And used lagged gene expression to predict the induction or repression state.}

\item{weight}{the final GRN is defined as (1-w)*Granger_coefficient + w*GENIE3_coefficient, if aggre_method = "manual"
Default: 0.2}

\item{redirected}{if re-estimate the causal direction through granger causal test.
Default: FALSE}

\item{fate_method}{user need to specific which fate probability they prefer to use, "cellrank" or "palantir"
Default: "palantir"}

\item{aggre_method}{use which aggregation method, "RobustRankAggreg", "SUMMA" or "manual". If use "manual", need to specific weight parameter.}

\item{restart}{if re-estimate the GRN. Default: FALSE}

\item{n.cores}{The number of cores. Default: 8}

\item{work_dir}{user could specific the working directory through input the address, and the inferred GRN would be saved in this directory.
Otherwise the local working directory would be used (getwd())
Default: NULL}

\item{GRN}{a list of cell-fate specific GRN}

\item{coef}{a list of granger causal correlation matrix}
}
\value{
a list contains following objects
}
\description{
This function is for generating the cell fate specific weighted gene regulatory network (GRN). Using the GRN skeleton and fate probability matrix, FateCausal would learn the granger causal using fate probability to order gene expression.
}
\examples{
\dontrun{
GRN.filter <- FateCausal(dyn.out,velo_infor = FALSE,L=80,alpha = 0,weight=0.8,lambda = 100,aggre_method = "RobustRankAggreg",restart = TRUE)
}

}
