% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FateDynamic.R
\name{FateDynamic}
\alias{FateDynamic}
\title{A end-to-end fate dynamic inference}
\usage{
FateDynamic(
  sce,
  global_params = list(),
  palantir_params = list(),
  cellrank_params = list(),
  work_dir = NULL
)
}
\arguments{
\item{sce}{A SingleCellExperiment object need to contain the "count" assay. "spliced" and "unspliced" assays are required if run cellrank}

\item{global_params}{a list object could manually set the core parameters of FateDynamic
see the details of parameters through ?check_global_params}

\item{palantir_params}{a list object could manually set the core parameters of palantir
see the details of parameters through ?check_palantir_params}

\item{cellrank_params}{a list object could manually set the core parameters of cellrank
see the details of parameters through ?check_cellrank_params}

\item{work_dir}{user could specific the working directory through input the address, and the varID object (pruneKnn) would be saved in this directory.
Otherwise the local working directory would be used (getwd())
Default: NULL}
}
\description{
A end-to-end cell dynamic inference. Run this function before RunNetID, if set dynamicInfor in RunNetID equal to TRUE.
}
