% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPlot.R
\name{graphPlot}
\alias{graphPlot}
\title{Plot Module Network}
\usage{
graphPlot(fate_grn, testRes, TF, lineage, sign = TRUE, cutoff = 0)
}
\arguments{
\item{fate_grn}{A list object containing the lineage-specific gene regulatory network (output of FateCausal).}

\item{testRes}{Module test result (output of the ModuleFinder).}

\item{TF}{The name of the transcription factor.}

\item{lineage}{The name of the lineage.}

\item{sign}{If visualizing the activation or inhibition regulation.}

\item{cutoff}{The minimum threshold of the network edges.}
}
\description{
Visualize the identified module.
}
\examples{
\dontrun{
graphPlot(grn_list, result, "Klf1", "Ery")
}

}
