% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleFinder.R
\name{ModuleFinder}
\alias{ModuleFinder}
\title{Module Identification}
\usage{
ModuleFinder(
  fate_grn,
  lineage,
  root_tf,
  method = "spinglass",
  steps = 30,
  gamma = 0.5,
  sign = TRUE,
  output_type = "list"
)
}
\arguments{
\item{fate_grn}{A list object containing the lineage-specific gene regulatory network (output of FateCausal).}

\item{lineage}{Specify the name of the lineage.}

\item{root_tf}{The transcription factors for which the module needs to be identified.}

\item{method}{Module identification method: "rw" (random walk) or "spinglass" (spinglass).
Default: spinglass.}

\item{steps}{Tunable parameter for random walk. Description: the number of steps to take.
Default: 30.}

\item{gamma}{Tunable parameter for spinglass. Description: Real constant representing the gamma argument of the algorithm. This balances the importance of present and non-present edges in a community. A community is a set of vertices having many edges within the community and few edges outside. The default value of 1.0 makes existing and non-existing links equally important. Smaller values emphasize existing links, while greater values emphasize missing links.
Default: 0.5.}

\item{sign}{Whether to consider activation or inhibition regulation.}

\item{ModuleList}{A list of identified modules.}

\item{ModuleTestResult}{A data.frame containing the statistical test results of the modules.}
}
\value{
A list containing the following objects:
}
\description{
Identify modules from the gene regulatory network.
}
\examples{
\dontrun{
result = ModuleFinder(fate_grn, "Ery", c("Gata1"), method = "spinglass")
}

}
