% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFateProb.R
\name{plotFateProb}
\alias{plotFateProb}
\title{Plot Cell Labels and Fate Probability on Dimensional Reduction Space}
\usage{
plotFateProb(
  dyn.out,
  basis,
  celltype = NULL,
  basis_name = "Dim",
  lineage = NULL
)
}
\arguments{
\item{dyn.out}{The output of RunNetID.}

\item{basis}{A 2D array with (cell*2) dimensions, where columns represent coordinates.}

\item{celltype}{Cell type labels.}

\item{basis_name}{The name of the dimensional reduction space.}

\item{lineage}{Lineage name.}
}
\description{
A simple function to plot cell fate probability and cell type labels.
}
\examples{
\dontrun{
## Plot cell types
plotFateProb(dyn.out, basis = dyn.out$basis, basis_name = "PCA", celltype = celltype)
## Plot fate probability
plotFateProb(dyn.out, basis = dyn.out$basis, basis_name = "PCA", lineage = colnames(dyn.out$fate_prob)[1])
}

}
