% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cellrank_params.R
\name{check_cellrank_params}
\alias{check_cellrank_params}
\title{A function to check NetID}
\usage{
check_cellrank_params(params)
}
\arguments{
\item{mode}{The velocity mode. Default: stochastic.}

\item{weight_connectivities}{Default: 0.2.}

\item{ncores}{Default: 6.}

\item{plot}{Default: TRUE.}

\item{tolerance}{Default: 1e-6.}
}
\description{
This function checks the parameters for the FateDynamic function.
FateDynamic provides cellrank_params, allowing users to specify desired parameters. The available parameters are listed as follows:
}
