% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_global_params.R
\name{check_global_params}
\alias{check_global_params}
\title{A check function for NetID}
\usage{
check_global_params(params)
}
\arguments{
\item{min_counts}{Default: 20}

\item{nhvgs}{default: 2000}

\item{npcs}{default: 30}

\item{knn}{default: 30}

\item{velo}{if run the velocity model (cellrank), default: TRUE}

\item{fate_predict}{default: TRUE}

\item{cluster_label}{specific the attributes in SingleCellExperiment object that contains the cluster label information}
}
\description{
a check function for FateDynamic function.
FateDynamic have provided global_params, user could specific the parameters the want, the parameters are listed as follow
}
