% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_netID_params.R
\name{check_netID_params}
\alias{check_netID_params}
\title{A Function to Validate NetID Parameters}
\usage{
check_netID_params(params)
}
\arguments{
\item{g}{Manually set the genes users want to analyze in their dataset. Default: NULL.}

\item{var}{If using variable genes to calculate principal components, used by the geosketch method. Default: FALSE.}

\item{do.prune}{If prune KNN graph.}

\item{SNN}{Use the Shared Nearest-Neighbor graph. Default: FALSE.}

\item{sampled_cells}{The barcode or ID of sampled cells.}

\item{sketch.method}{Perform sketching sampling on single-cell datasets: "random", "geosketch", or "SeuratSketching".}

\item{ndim}{Dimensions of PCs, used by the geosketch method. Default: 30.}

\item{n_cell}{The number of sampled cells. Default: 500.}

\item{Threshold_Num}{The minimum number of nearest neighbors for each seed cell after assignments.}

\item{normalize}{If performing normalization to the count matrix. Default: FALSE.}

\item{prior_net}{If a binary matrix indicates the prior knowledge of gene regulation, where rows are regulators and columns are targets.}
}
\description{
This function validates parameters for the RunNetID function.
RunNetID provides netID_params; users can specify the desired parameters, which are listed as follows:
}
