% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FateDynamic.R
\name{FateDynamic}
\alias{FateDynamic}
\title{An End-to-End Fate Dynamic Inference}
\usage{
FateDynamic(
  sce,
  global_params = list(),
  palantir_params = list(),
  cellrank_params = list(),
  work_dir = NULL
)
}
\arguments{
\item{sce}{A SingleCellExperiment object that needs to contain the "count" assay. "spliced" and "unspliced" assays are required if running cellrank.}

\item{global_params}{A list object that can manually set the core parameters of FateDynamic. Refer to the details of parameters using ?check_global_params.}

\item{palantir_params}{A list object that can manually set the core parameters of palantir. Refer to the details of parameters using ?check_palantir_params.}

\item{cellrank_params}{A list object that can manually set the core parameters of cellrank. Refer to the details of parameters using ?check_cellrank_params.}

\item{work_dir}{Users can specify the working directory by inputting the address, and the varID object (pruneKnn) will be saved in this directory. Otherwise, the local working directory will be used (getwd()).
Default: NULL.}
}
\description{
This function provides an end-to-end cell dynamic inference process. Run this function before RunNetID, if you set dynamicInfor in RunNetID equal to TRUE.
}
