% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNetID.R
\name{RunNetID}
\alias{RunNetID}
\title{Run NetID model}
\usage{
RunNetID(
  sce,
  min_counts = 10,
  varID_res = NULL,
  knn = 30,
  regulators = NULL,
  targets = NULL,
  netID_params = list(),
  velo = TRUE,
  dynamicInfer = TRUE,
  maxState = 5,
  cut_off = 2,
  work_dir = NULL
)
}
\arguments{
\item{sce}{A SingleCellExperiment object need to contain the "count" assay,
if it contains the "spliced" and "unspliced" assay, the NetID would run on "spliced" assay}

\item{min_counts}{minimum detected counts for genes, the genes which has counts > min_counts would be preserved
Default: 10}

\item{varID_res}{a varID object generated by pruneKnn}

\item{knn}{number of nearest neighbourhoods
Default: 30}

\item{regulators}{a gene list contains the feature names of regulators}

\item{targets}{a gene list contains the feature names of targets}

\item{netID_params}{a list object could manually set the parameter for the core algorithm of netID
see the details of parameters through ?check_netID_params}

\item{velo}{if use the fate probability inferred results from RNA velocity based method (cellrank), make sure you have run FateDynamic function
before you want to set this parameter = True
Default: FALSE}

\item{dynamicInfer}{if use the fate probability inferred results, make sure you have run FateDynamic function before you want to set this parameter = TRUE
Default: FALSE}

\item{maxState}{NetID would assign the cells to a cell fate (lineage) though clustering the cells according to the fate probability. Internally, NetID
would use Gaussain Mixture Model, and it would use BIC to choose optimal number of clusters. User need to set the maximum number of clusters.
Default: 5}

\item{cut_off}{Each cluster would assign to a cell fate through comparing the fate probability fold change (e.g. cell fate A vs. others), a cell fate with
fold change > cut_off would be assigned to the corresponding cluster
Default: 2}

\item{work_dir}{user could specific the working directory through input the address, and the varID object (pruneKnn) would be saved in this directory.
Otherwise the local working directory would be used (getwd())
Default: NULL}

\item{unique_assign}{if uniquely assign the cell fate to each cluster
Default: FALSE}

\item{restart}{RunNetID would automatically run the pruneKnn if it couldn't find the VarID_res.rds in the working directory. if VarID_res.rds already have
and user still want to re-run the pruneKnn, set the restart = TRUE
Default: FALSE}

\item{no_cores}{The number of cores would be used. Default: 4}

\item{fate_prob}{cell fate probability matrix learned by palantir, only be saved when dynamicInfer = TRUE}

\item{LineageClass}{the assigment of lineage type to each cell. The palantir fate probability matrix is used when perform assignment}

\item{fate_prob_velo}{cell fate probability matrix learned by palantir, only be saved when dynamicInfer = TRUE}

\item{LineageClass_velo}{the assigment of lineage type to each cell. The cellrank fate probability matrix is used when perform assignment}

\item{pseudotime}{the pseudotime for each cell. pseudotime is learned by palantir}

\item{velocity_pseudotime}{the pseudotime for each cell. pseudotime is learned by RNA velocity (scVelo)}

\item{GEP}{NetID inferred aggregated profile on sampling dataset, user could manually choose their GRN inference method and run on this expression profile}

\item{velo_m}{and velo_m_aggregate
the velocity matrix}

\item{skeleton}{a gene regulatory network skeleton (a binary matrix with 1 indicate have regulation)}

\item{var}{if using variable gene to calculate principal components, used by geosketch method, default: FALSE}

\item{sampled_cells}{the barcode or ID of sampled cells}

\item{sketch.method}{perform sketching sampling on single cell datasets, "geosketch" or "SeuratSketching"}

\item{ndim}{dimensions of PCs, used by geosketch method, Default: 30}

\item{n_cell}{the number of sampled cells, default: 500}

\item{Threshold_Num}{the minimum nn of each seed cells after assignments, default: 2}

\item{normalize}{if perform normalization to the count matrix, default: FALSE}

\item{prior_net}{if a binary matrix indicate the prior knowledge of gene regulation, row is the regulator, column is the target.}
}
\value{
a list contains following objects
}
\description{
This function is for generating the weighted gene regulatory network (GRN). NetID first sample the cells on the PCA space to make sure the sampled cell could cover the whole manifold. Then aggregate the neighbourhoods around each sampled cells on pruned KNN graph. And in the end run GENIE3 algorithm on the aggregated gene expression profile. The sampling and rational aggregation would make cells denoise and maintain the major variation, and speed up the GRN generating process.

a check function for RunNetID function.
RunNetID have provided netID_params, user could specific the parameters the want, the parameters are listed as follow
}
\examples{
\dontrun{
dyn.out <- RunNetID(sce,regulators = TF, targets = TF,netID_params = list(normalize=FALSE,n_cell = 5000),velo=FALSE,dynamicInfer = FALSE)
}

}
